Changed condition of setup of ant symlink
There are two tasks meant for setting up an ant symlink, one for Mac,
and one for other any other OS. Although both tasks have when statements
that specify to only run on MacOS, causing non Mac hosts to not have ant
set up properly. This change fixes the when statement on the task that
sets up a symlink on a non Mac host.

Signed-off-by: Colton Mills <millscolt3@gmail.com>