unixPB: Do not hardcode releases in apt (Debian)
Previously, when provisioning Debian/Ubuntu machines, apt sources for the
AdoptOpenJDK package repositories were hardcoded to the Debian release "jessy".
As a consequence, a newer Debian or any Ubuntu system got the wrong packages.
With this patch, Ansible will dynamically determine the name of the
Debian/Ubuntu release installed on the target host and correctly configure the
apt source so that Debian Buster will get packages for Buster and so forth.